+++ lineage/BaumWelch.py
@@ -86,8 +86,9 @@ def do_M_step(tHMMobj: list, MSD: list, betas: list, gammas: list):
     # If thmmObj[0] satisfies this "if", then all the objects in this population do.
     if tHMMobj[0].estimate.fT is None:
         assert tHMMobj[0].fT is None
-        T = do_M_T_step(tHMMobj, MSD, betas, gammas)
-        # T = np.ones((gammas[0][0].shape[1], gammas[0][0].shape[1])) / gammas[0][0].shape[1]
+        # T = do_M_T_step(tHMMobj, MSD, betas, gammas)
+        ##### the following line will replace line 89 in case we want to have equal transitions between all states.
+        T = np.ones((gammas[0][0].shape[1], gammas[0][0].shape[1])) / gammas[0][0].shape[1]
 
         # all the objects in the population have the same T
         for t in tHMMobj:
